Remove Duplicates from sorted Array
Used to remove duplicates from the sorted array , tried to achieve in O(1) extra Memory. The output will be the first part of array includes the distinct elements and remaining part of the array will be the duplicate elements of it. Thus returns the length as number of distinct element

e.g. 
i/p : [1,1,2]
o/p : [1,2,1] , length = 2

i/p : [0,0,1,1,1,2,2,3,3,4]
o/p : [0,1,2,3,4,0,1,2,3] , length = 5